integration folders refactor
