+++ src/test/java/co/mercadolibre/SocialMeli/integration/controller/SellerControllerIntegrationTest.java
@@ -1,9 +1,8 @@
-package co.mercadolibre.SocialMeli.Integration.controller;
+package co.mercadolibre.SocialMeli.integration.controller;
 
 import co.mercadolibre.SocialMeli.dto.response.ClientFollowedDTO;
-import co.mercadolibre.SocialMeli.entity.User;
+import co.mercadolibre.SocialMeli.dto.response.SellerFollowersDTO;
 import co.mercadolibre.SocialMeli.util.IntegrationData;
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -16,16 +15,13 @@ import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
-import java.util.List;
-import java.util.Set;
-
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SpringBootTest
 @AutoConfigureMockMvc
-public class UserControllerIntegrationTest {
+public class SellerControllerIntegrationTest {
 
     @Autowired
     MockMvc mockMvc;
@@ -33,24 +29,24 @@ public class UserControllerIntegrationTest {
     ObjectMapper objectMapper = new ObjectMapper();
 
     @Test
-    @DisplayName("T0003 - Lista seguidos - Camino Bueno")
-    public void listFollowedGoodTest() throws Exception {
+    @DisplayName("T0003 - Lista seguidores - Camino Bueno")
+    public void listFollowersGoodTest() throws Exception {
 
         int userId = 1;
 
-        MvcResult response = this.mockMvc.perform(MockMvcRequestBuilders.get("/users/{userId}/followed/list",userId))
+        MvcResult response = this.mockMvc.perform(MockMvcRequestBuilders.get("/users/{userId}/followers/list",userId))
                 .andDo(print()).andExpect(status().isOk())
                 .andExpect(content().contentType("application/json"))
                 .andReturn();
 
-        ClientFollowedDTO followedResponse = objectMapper.readValue(response.getResponse().getContentAsString(), ClientFollowedDTO.class);
-        Assertions.assertEquals(IntegrationData.getClientFollowedDTO(), followedResponse);
+        SellerFollowersDTO followersResponse = objectMapper.readValue(response.getResponse().getContentAsString(), SellerFollowersDTO.class);
+        Assertions.assertEquals(IntegrationData.getListFollowersDTO(), followersResponse);
 
     }
 
     @Test
-    @DisplayName("T0003 - Lista seguidos - Camino malo, usuario no encontrado")
-    public void listFollowedNotFoundSellerTest() throws Exception {
+    @DisplayName("T0003 - Lista seguidores - Camino malo, vendedor no encontrado")
+    public void listFollowersNotFoundSellerTest() throws Exception {
 
         int userId = 10;
 
@@ -59,6 +55,20 @@ public class UserControllerIntegrationTest {
                 .andExpect(content().contentType("application/json"))
                 .andExpect(MockMvcResultMatchers.jsonPath("$.message")
                         .value("El usuario con el id %d no se ha encontrado".formatted(userId)));
+
+    }
+
+    @Test
+    @DisplayName("T0003 - Lista seguidores - Camino malo, no es vendedor")
+    public void listFollowersNotASellerTest() throws Exception {
+
+        int userId = 2;
+
+        this.mockMvc.perform(MockMvcRequestBuilders.get("/users/{userId}/followers/list",userId))
+                .andDo(print()).andExpect(status().isNotFound())
+                .andExpect(content().contentType("application/json"))
+                .andExpect(MockMvcResultMatchers.jsonPath("$.message")
+                        .value("El usuario con el id %d no es un vendedor".formatted(userId)));
 
     }
